Modify data model from post to level
