+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 .vscode
 *.log
 .env*
+src/generated/
