+++ database/prisma.yml
@@ -1,5 +1,5 @@
 # the name for the service (will be part of the service's HTTP endpoint)
-service: authplate
+service: emm-backend
 
 # the cluster and stage the service is deployed to
 stage: ${env:PRISMA_STAGE}
