+++ database/seed.graphql
@@ -2,21 +2,25 @@ mutation {
   createUser(data: {
     email: "developer@example.com"
     password: "$2a$10$hACwQ5/HQI6FhbIISOUVeusy3sKyUDhSq36fF5d/54aAdiygJPFzm" # plaintext password: "nooneknows"
-    name: "Sarah"
-    posts: {
-      create: [{
-        title: "Hello World"
-        text: "This is my first blog post ever!"
-        isPublished: true
-      }, {
-        title: "My Second Post"
-        text: "My first post was good, but this one is better!"
-        isPublished: true
-      }, {
-        title: "Solving World Hunger"
-        text: "This is a draft..."
-        isPublished: false
-      }]
+    name: "Morio"
+    levels: {
+      create: [
+        {
+          name: "Yoshi's Island"
+          description: "This is my first level ever!"
+          isPublished: true
+        },
+        {
+          name: "Donut Plains"
+          description: "My first level was good, but this one is better!"
+          isPublished: true
+        },
+        {
+          name: "Mango Madness"
+          description: "This is a new secret level."
+          isPublished: false
+        }
+      ]
     }
   }) {
     id
