+++ src/resolvers/Query.js
@@ -2,7 +2,7 @@ const { getUserId } = require('../utils')
 
 const Query = {
   feed(parent, args, ctx, info) {
-    return ctx.db.query.posts({ where: { isPublished: true } }, info)
+    return ctx.db.query.levels({ where: { isPublished: true } }, info)
   },
 
   drafts(parent, args, ctx, info) {
@@ -10,16 +10,16 @@ const Query = {
 
     const where = {
       isPublished: false,
-      author: {
+      creator: {
         id
       }
     }
 
-    return ctx.db.query.posts({ where }, info)
+    return ctx.db.query.levels({ where }, info)
   },
 
-  post(parent, { id }, ctx, info) {
-    return ctx.db.query.post({ where: { id } }, info)
+  level(parent, { id }, ctx, info) {
+    return ctx.db.query.level({ where: { id } }, info)
   },
 
   me(parent, args, ctx, info) {
