+++ src/resolvers/index.js
@@ -1,14 +1,14 @@
 const { Query } = require('./Query')
 const { Subscription } = require('./Subscription')
 const { auth } = require('./Mutation/auth')
-const { post } = require('./Mutation/post')
+const { level } = require('./Mutation/level')
 const { AuthPayload } = require('./AuthPayload')
 
 module.exports = {
   Query,
   Mutation: {
     ...auth,
-    ...post,
+    ...level,
   },
   Subscription,
   AuthPayload,
