+++ src/schema.graphql
@@ -1,22 +1,22 @@
-# import Post from "./generated/prisma.graphql"
+# import Level from "./generated/prisma.graphql"
 
 type Query {
-  feed: [Post!]!
-  drafts: [Post!]!
-  post(id: ID!): Post
+  feed: [Level!]!
+  drafts: [Level!]!
+  level(id: ID!): Level
   me: User
 }
 
 type Mutation {
   signup(email: String!, password: String!, name: String!): AuthPayload!
   login(email: String!, password: String!): AuthPayload!
-  createDraft(title: String!, text: String!): Post!
-  publish(id: ID!): Post!
-  deletePost(id: ID!): Post!
+  createDraft(title: String!, text: String!): Level!
+  publish(id: ID!): Level!
+  deleteLevel(id: ID!): Level!
 }
 
 type Subscription {
-  feedSubscription: PostSubscriptionPayload!
+  feedSubscription: LevelSubscriptionPayload!
 }
 
 type AuthPayload {
@@ -28,5 +28,5 @@ type User {
   id: ID!
   email: String!
   name: String!
-  posts: [Post!]!
+  levels: [Level!]!
 }
